Various correctness improvements

+++ dex/engine-api/src/main/java/org/dependencytrack/dex/engine/api/WorkflowRun.java
@@ -34,7 +34,7 @@ public record WorkflowRun(
         int workflowVersion,
         WorkflowRunStatus status,
         @Nullable String customStatus,
-        short priority,
+        int priority,
         @Nullable String concurrencyGroupId,
         @Nullable Map<String, String> labels,
         Instant createdAt,
